Added required and validation on text fields
removed css files , will create a css file later on

+++ src/main/java/com/bitvault/ui/views/categories/CategoryVM.java
@@ -4,17 +4,22 @@ import com.bitvault.services.interfaces.ICategoryService;
 import com.bitvault.ui.async.AsyncTask;
 import com.bitvault.ui.model.Category;
 import com.bitvault.util.Result;
+import javafx.application.Platform;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 public class CategoryVM {
 
     private final ICategoryService categoryService;
     private final SimpleBooleanProperty loading = new SimpleBooleanProperty();
     private final ObservableList<CategoryRowView> categories = FXCollections.observableArrayList();
+
+    private List<Category> oldCategories;
 
     public CategoryVM(ICategoryService categoryService) {
         this.categoryService = categoryService;
@@ -29,9 +34,11 @@ public class CategoryVM {
     }
 
     private void onLoaded(List<Category> categories) {
+
+        this.oldCategories = new ArrayList<>(categories);
         final List<CategoryRowView> categoryRowViews = categories
                 .stream()
-                .map(category -> CategoryRowView.createFromCategory(category, this::onDelete))
+                .map(category -> CategoryRowView.createFromCategory(category, this::onDelete, this::onSave))
                 .toList();
         this.categories.addAll(categoryRowViews);
         this.loading.set(false);
@@ -50,16 +57,62 @@ public class CategoryVM {
         return categoriesResult.get();
     }
 
-    private void onDelete(CategoryRowView categoryRowView) {
+    private boolean onDelete(CategoryRowView categoryRowView) {
+
+        Optional<Category> foundInOld = oldCategories.stream()
+                .filter(category -> category.id().equals(categoryRowView.getUniqueId()))
+                .findAny();
+
+        //if is not in old list then we can remove without checking service
+        if (foundInOld.isEmpty()) {
+            Platform.runLater(() -> this.categories.remove(categoryRowView));
+            return true;
+        }
+
         Result<Boolean> booleanResult = categoryService.delete(categoryRowView.getUniqueId());
         if (booleanResult.isFail()) {
-            return;
+            return false;
         }
-        this.categories.remove(categoryRowView);
+        Platform.runLater(() -> this.categories.remove(categoryRowView));
+        return true;
+    }
+
+    private Result<String> onSave(CategoryRowView categoryRowView) {
+
+        Optional<Category> foundInOld = oldCategories.stream()
+                .filter(category -> category.id().equals(categoryRowView.getUniqueId()))
+                .findFirst();
+
+        //for update
+        if (foundInOld.isPresent()) {
+            Category updateCat = Category.createUpdate(foundInOld.get().id(), categoryRowView.getCategoryName(), categoryRowView.getColor(), "Password");
+            Result<Category> updateResult = categoryService.update(updateCat);
+            if (updateResult.isFail()) {
+                updateResult.getError().printStackTrace();
+                //TODO
+                return Result.error(updateResult.getError());
+            }
+            return Result.ok(updateResult.get().id());
+        }
+
+        //for new
+        Category category = Category.createNew(categoryRowView.getCategoryName(), categoryRowView.getColor(), "Password");
+        Result<Category> categoryResult = categoryService.create(category);
+
+        if (categoryResult.isFail()) {
+            categoryResult.getError().printStackTrace();
+            //TODO
+
+            return Result.error(categoryResult.getError());
+        }
+
+        oldCategories.add(categoryResult.get());
+        return Result.ok(categoryResult.get().id());
+
     }
 
     public CategoryRowView addNewCategory() {
-        CategoryRowView aNew = CategoryRowView.createNew(this::onDelete);
+        CategoryRowView aNew = CategoryRowView.createNew(this::onDelete, this::onSave);
         this.categories.add(aNew);
         return aNew;
     }
